+++ pythainlp/transliterate/core.py
@@ -115,6 +115,8 @@ def transliterate(
         * *tltk_ipa* - tltk, output is International Phonetic Alphabet (IPA)
         * *thaig2p_v2* - Thai Grapheme-to-Phoneme,
           output is IPA. https://huggingface.co/pythainlp/thaig2p-v2.0
+        * *thaig2p_v3* - Thai Grapheme-to-Phoneme using ONNX model (v3),
+          output is IPA. https://github.com/wannaphong/thai-g2p-v3
         * *umt5_thaig2p* - Thai Grapheme-to-Phoneme,
           output is IPA, powered by UMT5.\
           https://huggingface.co/B-K/umt5-thai-g2p-v2-0.5k
@@ -158,6 +160,8 @@ def transliterate(
         from pythainlp.transliterate.iso_11940 import transliterate  # type: ignore[assignment]  # noqa: I001
     elif engine == "thaig2p_v2":
         from pythainlp.transliterate.thaig2p_v2 import transliterate  # noqa: I001
+    elif engine == "thaig2p_v3":
+        from pythainlp.transliterate.thaig2p_v3 import transliterate  # noqa: I001
     elif engine == "umt5_thaig2p":
         from pythainlp.transliterate.umt5_thaig2p import transliterate  # noqa: I001
     else:  # use default engine: "thaig2p"